MXS-4872 Fix VirtualScrollTbl contextmenu handler
If a slot is used, `event.target` will be the slot content instead of
the parent element. Therefore, the activatorID can't not be retrieved.

The activatorID can be retrieved within the default slot of `GblTooltipActivator`,
so it can be passed to the `contextmenu` handler.

This also fixes the binding of events in `HistoryAndSnippetCtr` component.