MXS-5032: Store simple PS command as COM_QUERY
If a binary protocol prepared statement is executed that takes no
parameters and that refers to a COM_STMT_PREPARE that's still in the
history, a COM_QUERY version of the COM_STMT_PREPARE is stored in the
history instead. This will have the same effect as executing a
COM_STMT_EXECUTE with an ID of -1 would have except that the COM_QUERY
will still take effect even after a COM_STMT_CLOSE.

Documented the new behavior in the `prune_sescmd_history` parameter as
it is what controls how the history handling is done. A separate section
in the documentation that describes the session commands in detail would
be good.