Adjust readwritesplit error messages
The errors that are propagated upwards from a single-level
readwritesplit service in 24.02 are not informative enough on their
own. For example, if a connection is lost while a transaction is open,
the following was the only error that was logged.

Cannot replay transaction because 'transaction_replay' is not enabled.

This was lacking context which made it hard to determine exactly why
this error was logged. Always prepending the connection name and the
original error will show the full "error stacktrace" to the end user
which should make it a lot easier to reason about the origins of
failures when there are multiple services.

Adjusted the error messages in the code called by handleError to be more
readable and consistent. Most of the errors now have a "closing session"
message to indicate that this is a fatal error that will close the
session.