Use IF NOT EXISTS instead of OR REPLACE
In testing, the CREATE OR REPLACE appears to be executed twice if the
original statement was cut short by the network being blocked. The
second execution then later works and gets binlogged with the expected
GTID. However, the original failed CREATE appears later on with a higher
GTID but the original thread_id value which then causes the table to be
empty again and the SELECT will not match any rows.

Using IF NOT EXISTS will work in this case and it being executed twice
will not be a problem.