Add Ubuntu Jammy ASAN workaround
Some Ubuntu Jammy images come with kernels that have the high-entropy
ASLR enabled. This causes conflicts with the ASAN address space which
makes tests fail randomly. This is caused by a bug in ASAN that was
later fixed in Clang but the fixed ASAN code does not appear to be in
GCC 13 that's on Ubuntu Jammy.

A workaround to this is to enable the low-entropy ASLR by setting the
randomization size to 28 bits.