MXS-5048: Wildcard hosts must be treated as HOSTNAME
If the host pattern contains wildcards, it must be treated as a HOSTNAME
type. Otherwise a hostname pattern with a prefix that matches both an
address and a hostname gets wrongly classified as an address.

The exceptions to this are addresses that start with something that
looks like an IPv4 address. MariaDB rejects matching against any
hostnames that look like the start of an IPv4 address. MaxScale now
behaves similarly to MariaDB and also logs a warning if a IP address
resolves to a hostname that looks like an IPv4 address.