MXS-5038: Fix BINARY field conversion to JSON
The BINARY type fields were not processed correctly and rows with them
ended up being skipped if the data wasn't valid UTF8.

There was also an inconsistency in how the RowEventHandler API was
called for fixed length string and binary types and variable length
string and binary types. The former were always treated as strings
whereas the latter was treated as a string only for true string
types. Unfortunately, the only way to detect the difference is to
compare the field type in the SQL.

With the proper API usage, the type of BINARY must now be `bytes`
instead of `string`, otherwise the Avro C API calls will fail.