MXS-4871 Delay 300ms before showing the tooltip
Having a reasonable delay would improve the UX.
300ms would be long enough to ensure user intent to
see the tooltip content.

This adjusts the usage of `VTooltip` in some places so that
the tooltip is activated based on the mouseover event instead of
using `modelValue`. This change is necessary because when using
`modelValue`, the `open-delay` prop is ignored.

This also improves `GblTooltipActivator` to cancel the debounce
function on mouseleave.