MXS-5039 Set session-level wait_timeout when cooperative monitoring is on
This causes the server to close the monitor connection if it has
been idle for longer than the timeout. Setting a reasonable
wait_timeout for monitor connections releases stale locks
even if tcp keepalive settings are not used. The timeout is
altered when starting a cluster operation to ensure connections
are not lost during the operation.