MXS-4871 Fix `on-tree-changes` event
After expanding a node, it failed to retrieve new completion items
because the `on-tree-changes` event was not emitted.

To resolve this issue, the `tree` property must be assigned a
completely new object instead of simply referencing the `items`
proxy object. Otherwise, its watcher won't be able to detect
changes, as the old and new values will be identical due to their
shared object reference with the `items` proxy.