MXS-4871 Place `workspace` composables into `services` dir
`services` is a commonly used term to denote a layer that contains
business logic. All ORM actions or api calls in the `workspace` view
should be placed in this dir to enhance the scalability and
maintainability of the application.

This also refactors those composable hooks into pure functions as
they are service functions that could be called in non-vue components.