MXS-5063: Don't trigger fake events during throttling
If a fake read event is triggered while reads have been disabled, the
throttling would be bypassed. By storing EPOLLIN in
m_triggered_event_old in this case, the enabling of reads will then
handle it the same way it would've handled it if a fake EPOLLIN event
was pending when the reads were originally disabled.