Fix debug assertion on invalid TLS key
The configuration validation in config.cc behaved in a way where it
expected the values to be available even if they weren't valid.

This happened in maxctrl_basic for an unknown reason but the added test
case reproduces the problem.

Also adjusted the debug assertion that was hit to also print the
contents of the JSON if it wasn't valid.