MXS-5068: Fix users_refresh_time=0s
The throttling won't be ever enabled if the refresh interval is set to
zero. With this, the behavior is as documented.

Added a test case that reproduces the original problem and verifies that
the fix works as expected.

This should also fix the auth_wchost that happened to fail due to the
hard-coded limit of 5 refreshes being exceeded by the addition of the
new test case.