MXS-5021: Use GDB for stacktraces if available
Before 23.08, this is always the best method for capturing stacktraces
as it gets fully demangled names for all functions and for all
threads. In 23.08 this is less important as it uses the internal
profiler function for generating the stacktrace but even in this case
GDB output is slightly more accurate.