MXS-5027: Add `allow_duplicates` to schemarouter
This will be a more convenient way to disable it that also skips the
regular expression execution. This is also more efficient as deployments
with tens of thousands of tables on different shards would go through
all of the values even if all duplicates would be ignored.

Modifies the configuration of the sharding test to use the new
option. This way the value is used in a way that tests that it works.