MXS-4871 Emit node-hovered event from `VirSchemaTree`
`VirSchemaTree` was utilizing the Vuetify VHover component to add
mouse events for detecting whether the node is hovered or not.
However, this occasionally conflicted with the `mouseover` event
in the `SchemaTreeCtr` component. For the sake of simplicity,
`VirSchemaTree` will now be responsible for emitting the
`node-hovered` event.

This also migrate the unit tests of `SchemaTreeCtr`, some of the tests
are not migrated as they are now obsolete.