MXS-5092: Remove leading signs from canonicals
The sign can be omitted if the first non-space character that comes
before it is an arithmetic, logical or comparison operator or if it's
either an opening parenthesis or a comma. In these cases the SQL syntax
does not allow a sign to be used to mean subtraction or addition.

Adjusted the test cases to expect the new behavior. Also changed the old
canonical implementation to do the same sign elimination for the sake of
making the tests pass.