MXS-4871 Fix ERD diagram context menu visibility
After opening the context menu, it can be reopened by left-clicking on
the graph board, which is not the expected behavior. This was previously
functioning correctly, but due to changes in the `CtxMenu` component
during the migration, it is now broken.

This adds a watcher for `showCtxMenu` and resets `activeCtxItem` to
null to prevent the context menu from opening with a left-click action.