MXS-5083: Make ssl_version work like tls_version
The ssl_version is now an enumumeration mask that accepts a combination
of the values. The default value of MAX still has the same behavior that
it had previously and single values again behave like they did in legacy
versions. Multiple values now behave the same way that the MariaDB
tls_version parameter does.