Rewrite mxs822_maxpasswd
The test was broken as it did not detect a failure to decrypt the
password. It also did not create a connection even though the comments
claimed it did.

To correctly detect whether an encrypted password worked, a connection
needs to be created to the backend. Additionally, the test now checks
that the decryption results in the same password that was originally
passed to maxpasswd.

The sleeps in the test are not necessary and removing them speeds up the
testing significantly, from about 20 seconds to about 4 seconds.