Add support for manual gcov builds
If the MXS_SKIP_BUILD=Y environment variable is set, the gcov_start test
will skip the build part. Combined with MXS_BUILD_ROOT, this can be used
to manually start the gcov collection.

The gcov_stop test can then be used to collect the results. The test
also now prints the location of the file and when run from a terminal
that supports links, clicking it will open it in a browser.