MXS-4871 Fix editor caching issue
When the user deletes the last query tab, if the last query tab is
not of the `SQL_EDITOR` type but of the `ALTER_EDITOR` or
`INSIGHT_VIEWER` type, the built-in `KeepAlive` component in Vue 3
will fail to handle the changes in the editor type. Therefore,
the query tab's type must be used along with the query tab ID as
the key for the `KeepAlive` component.