MXS-3852 Add process types filter
Users can now filter the process list result set by either
`WORKSPACE` or `OTHERS` types. By default, all rows will
be shown.

This also adjusts `ConnDlg` to store the MaxScale object type
in the metadata which will be used to fetch the session object
if the connection is created from a listener or a service.