MXS-5104 Fix connection busy error while connecting to a listener
When connecting to a listener in the Query Editor, the following
error is shown: "ID 80d10003-6171-4dbb-8c8e-1c14a49f2153 is busy.
Connection busy, please wait."

This issue does not occur when connecting directly to a server
as the result is returned relatively fast. The problem stems
from the fact that after opening the query tab connection,
the connection object is inserted immediately to the model,
triggering the watcher in the QueryEditor component. As
as result, there are sometimes concurrent queries.