MXS-4911 Add initial version of convert utility
The convert utility reads a mysqltest test-file, classifies all
statements and saves the information in a file one json-object
per line. The json-objects look like:

{
    "statement": "...",
    "sql_mode": "...",
    "classification": { ... }
}

The 'statement' contains the statement in question, 'sql_mode' the
mode the parser was in, and 'classification' the classification
result.

Functionality in 'compare' that was needed by 'convert' was factored
out into a common include file 'util.hh' and 'compare' modified
accordingly.