MXS-5106: Use capabilities of the available servers
If a server is not available and has not been contacted, the protocol
must be assumed to be the lowest known supported version. For example
with three 10.11 MariaDB servers out of which one is down, the assumed
version must still be that of MariaDB 10.11. Otherwise a single downed
node downgrades the protocol version of the whole cluster.

Any protocol incompatibilities will be resolved during connection
creation which will permanently close any unavailable backends. This may
cause interruptions but clients that expected an older protocol version
will never end up in this situation.

Extended the bulk_insert test case to cover this scenario.