MXS-5107 Fix `connection busy` error causing empty schema tree explorer
When the user runs a query and navigates away from the Query Editor
page, then returns to it, an error message appears:
'ID bfb44f7b-e900-4d83-97d6-24c7ea4e1ad9 is busy. Connection busy,
please wait.' Consequently, the schema tree explorer is empty.

The queries to populate the schema tree shouldn't be refetched
while a query is already running. This happens because the schema
tree data is cleared whenever re-entering the query editor page.

This checks for the existences of `tmp` models before initializing
new ones. However, this solution works only if the user doesn't
intentionally refresh the browser, as the schema tree data is stored
in memory. If the tree data is empty, it will still automatically
trigger actions to query the tree data.