MXS-3852 Fix duplicated Processlist query
The default value for `is_validating_conn` should always be true
whenever the workspace view is mounted. However, when the workspace
view is mounted a second time (i.e., when the user navigates away
from the workspace and then navigates back), the value for
`is_validating_conn` changes from false to true and then back to false.
This triggers the `queryEditorService.initialFetch` call twice.

This replaces the mentioned global state with a component state to
ensure that the initial value is always set to true.