MXS-4871 Fix odd behavior of virtual table column widths
When entering the workspace, the width is calculated while the app
sidebar is still expanded. When the sidebar collapses, the Vuetify
`v-resize` directive no longer triggers the resize change event as it
used to be in vuetify version 2. As a result, the workspace width
is calculated incorrectly, causing odd behavior in the virtual table
column widths.

This adds a custom resize observer directive with a debounced value
of 200ms to mimic the old behavior of the Vuetify `v-resize` directive.