MXS-3852 Adjust dynamic tab components rendering
This adjustment is required to enable the `VMenu` component in the
`SessionsTable` component to correctly calculate the position of the
activator. Without this, the menu tooltip will always appear at the
top left corner of the window.

Additionally, by doing this adjustment, if a tab is inactive,
it will now be destroyed which helps to reduce DOM population.