MXS-5105: Add mxb::lru_cache
The class impliements a LRU cache with no automatic eviction. This makes
it possible to do the cache eviction outside of the container based on
whatever criteria are needed.

Since most of the LRU use-cases will use the entry size as the eviction
criteria, there is a convenience constant that can be used to calculate
the estimated static overhead of one entry.

Added a unit test that tests the LRU behavior and makes sure that the
keys do not point to stale values.