MXS-5105: Use mxb::lru_cache in QCInfoCache
Minimal changes were required as the mxb::lru_cache behaves very close
to how a std::unordered_map behaves. The estimated size of the entry
uses the mxb::lru_cache estimate which itself was based on the 32 byte
extra overhead that was previously used. The current overhead is roughly
double of the previous one at 64 bytes per entry.