MXS-5049: Allow caching of hostname lookup results
The mxb::reverse_name_lookup() function can now optionally use cached
results. The cache is local to each thread and the entries are
invalidated after a hard-coded limit of 300 seconds.

DNS records have lifetimes from a handful of minutes up to about 30
minutes. 300 seconds seems to be a common low-end TTL value so using it
as the hard-coded value is a good starting point. Ideally the TTL value
would be from the DNS server but this would probably require using more
than just a call to getnameinfo().

Functions for directly accessing the thread-local cache are exposed in
the host.hh header. This will be needed by the client protocol code
where the hostname lookups are done asynchronously.