MXS-5049: Add host_cache_size
The parameter controls how many entries are stored in the cache for each
thread. The default value of 128 is the same that is in MariaDB.

With 128 entries, each a hostname that's 1024 bytes long, the worst-case
memory consumption per thread is about 160KiB. This is quite a small
amount of memory for modern systems and with a more realistic hostname
length of 255 bytes, this is under 62KiB per thread.

The REST-API as well as the MariaDB client protocol both use this cache.