MXS-5105: Use mxb::lru_cache in the REST-API
The user authentication has to hash the plaintext passwords into SHA256
which previously used a std::unordered_map as the cache.

The old cache eviction was slightly problemtic in that it assumed all
buckets were empty which theoretically made it "leak" memory if the
random bucket selected for eviction was always empty. A similar problem
was already found in the query classifier cache where the bucket
eviction used to walk the buckets until it found a non-empty one.