MXS-5037: Track reads and writes at the server level
The same logic that was used for the query latency histograms is now
used for the packet count tracking. All queries routed to a server
without the SERVER_MASTER status will be treated as reads and the query
classifier is only consulted for queries routed to a server with
SERVER_MASTER. Services also track the number of reads and writes done
through them but as the call to status() is recursive, the SERVER_MASTER
bit is not used there. Instead, the task of determining the packet type
is always delegated to the new get_packet_type() function.

Inverting the logic of the packet type computation in ServerEndpoint
keeps the amount of calls into the query classifier the same as it was
before for a single level service for all reads that are routed to a
slave server. This happens because the code in the ServerEndpoint now
first checks the status of the server before consulting the query
classifier which offsets the extra call that the service now
does. Additionally, the services will not call the status() function if
the session has the RCAP_TYPE_QUERY_CLASSIFICATION capability and
instead directly consult the query classifier. This avoids the recursive
status() call for services that would require query classification.

Making the Server and Service classes `final` also allows the calling
code in ServerEndpoint to avoid all the overhead of virtualization of
the status() function. For the Server class this is probably more
significant as it allows direct access to the m_status member.

Added a small test case into sanity_check that makes sure that the
packet counts are incremented as expected.