MXS-5041: Don't retry autocommit writes in trx replay
If transaction_replay is enabled with transaction_replay_safe_commit,
the functionality that delayed_retry normally enables is disabled when
there is no transaction open. This means that writes done with
autocommit are never retried as they're not safe to replay due to their
non-transactional behavior.

Extended the mxs4505_trx_safe_replay test case to also check that the
new behavior works. It also confirms that previously the autocommit
statements indeed were replayed.