Remove versioned links from the REST-API
The links in the static page served by the REST-API when HTTPS is not
enabled had links to documentation from MaxScale 2.4. Replaced the
versioned links to the parameters and tutorials and replaced it with a
link to the main MaxScale documentation page where the documentation for
all the versions is.