MXS-5112: Add PsReuse filter
The filter implements, and slightly extends, the functionality in
readwritesplit that `reuse_prepared_statements` enabled. The feature in
readwritesplit was untested and it did not work with direct PS
execution.

Moving the functionality into a filter will make it possible to still
enable it but without always having to pay the (insignificantly small)
cost of having in readwritesplit. The main benefit of this is
architectural clarity and the ability to use it with other routers.