Make mxs1245_batch_query faster
Instead of using the SLEEP() function, FLUSH TABLES WITH READ LOCK can
be used to prevent the queries from executing. This way the test can
proceed with the reading of the results as soon as the server has failed
which speeds up the test by about 15 seconds from the old runtime of 30
seconds.