MXS-5112: Add optimistictrx filter
The filter implements the optimistic_trx feature in readwritesplit in a
way that requires no transaction replay to work. The major difference is
that it uses START TRANSACTION READ ONLY to start an explicit read-only
transaction that when combined with readwritesplit, will behave
identically to the old feature in that it will be routed to a
slave. When the filter detects that a write is about to happen, it
issues a ROLLBACK and replays the start of the transaction
asynchronously. If the checksums do not match, the filter returns false
from clientReply which terminates the session.

The only major difference in the implementations is the checksum
algorithm and how it's computed. In readwritesplit, as the original
statements are executed as-is, the checksum can be computed over the raw
data. As the START TRANSACTION READ ONLY causes one bit in the OK and
EOF packets to differ, the checksum in the optimistictrx filter must
only be done on the data that's exposed in the mxs::Reply. In practice
the accuracy of this checksum should be the same as the other bits in
the status field are largely irrelevant.

Changed the tests to use the optimistictrx filter instead of the
optimistic_trx feature in readwritesplit. This is in preparation for
removing the feature from readwritesplit.