MXS-5016 Replace mxb::Json with bsoncxx
The original idea was to convert the document in the database
to a mxb::Json on which all manipulation would be made and then
at the end convert it to bson.

That would work fine as long as the accessed types are simple
ones, such as int, double or string, but would become quite
difficult if the types are such that appear to be regular Json
object, but that once translated into Bson are specific types.

So, better to just bite the bullet, turn the json from the
database into bson and manipulate that. A particular problem
is that in bsoncxx, a value cannot exist independently of a
containing document or array. That's the reason why bsoncxx
values are created into a builder.