MXS-5121: Query MariaDB 10.10 tables for charset byte
The character set byte that's sent to the client must be queried
differently in MariaDB 10.10 if collation_server is set to one of the
new UCA 14 collations. This is because the information_schema.COLLATIONS
table has NULL IDs in it and the value of @@global.collation_server is
no longer found there.

The new way to query the table would be to use the
information_schema.COLLATION_CHARACTER_SET_APPLICABILITY table where the
collation name is stored in its full form.

The ID values for these new collation go above 255 which means they
cannot be used as-is in the protocol. A mapping to the general_ci
collations must be done to make MaxScale behave identically to MariaDB.

The "collation" is in reality the character set that the server uses
even if the MariaDB network protocol documentation states that it's the
collation. The ID is, however, still stored in the
information_schema.COLLATIONS table so there is some confusion when it
comes to the naming of things.