MXS-5016 Work around SingleCommand assumptions
SingleCommand is not optimal as a base for Aggregate, as it
assumes SQL will always have to be executed. That is not the
case for Aggregate as it, when used in 'explain' mode, will
generate the response immediately. It's fairly close so now
prepare() is executed explicitly to find out whether we
need to explain or execute and then double preparation is
simply prevented.