MXS-5126: Fix CachePT diagnostics
Calling `maxctrl show filters` when the cache filter is configured and
populated with the thread-local cached, the call to Cache::do_get_info()
eventually ended up accessing data that was only available on a
RoutingWorker. Since the call is done on the MainWorker, the execution
must be moved temporarily to the first RoutingWorker.

Added a test case into maxctrl_basic that reproduces the crash and
verifies that the fix works.