MXS-4783: Ignore failing uservar setup in cache
If a session uses more than one cache filter session, the installation
of the @MAXSCALE.var user variable handlers for the lower level ones
will fail. This is currently expected as the variable handlers are
unique across the whole session.

Documented this limitation and changed the errors into info level
messages and remove the debug assertions. Also added a debug assertion
into Session::add_variable() to detect the case where the prefix is
missing.