MXS-4783: Only hold weak references to Routables
The ServiceEndpoint should be the only one that holds a
shared_ptr<Routable> that is not generated from a
weak_ptr<Routable>. Otherwise the call to ServiceEndpoint::close() will
not destroy the router session in the close() call like is expected.

The closing of the Routable could happen at a later point if a reference
to it would outlive the ServiceEndpoint::close() call but this would
complicate things when it comes to routing. If a Routable outlives the
ServiceEndpoint::close() call, subsequent calls to e.g. clientReply may
work part of the way up the call chain but they may encounter a
"partially closed" ServiceEndpoint that has been closed but for which
the Routable is still alive due to a shared_ptr<Routable> being alive
somewhere.