Fix race condition in mxb::latch
The notify_all() call must be made with the mutex locked to prevent a
race condition where a thread waiting for the latch to count down does
not see the notification sent by the thread that decrements the counter
to zero.

This was seen in a run of test_latch where the test timed out due to a
thread not being woken up. Running about 100 instances of the test in
parallel seems to reproduce it. With this commit, the problem is no
longer there and 10 iterations of 100 parallel test runs completed
without a problem.