MXS-5132: Always expect a proxy header
This will cause the correct error messages to be produced.

There was also a bug in the error generation where the sequence number
was set to 1 when it should be 2 for errors generated by the proxy
protocol processing stage.

Now with these two problems fixed, a MaxScale that connects to another
MaxScale will receive an error similar to the following:

  Monitor was unable to connect to server server1[127.0.0.1:4006] :
  'Proxy header is not accepted from ::ffff:127.0.0.1'