Make log_throttling faster and more stable
Running something CPU heavy while running the test_logthrottling unit
test easily caused it to fail. Since the unit testing isn't intended to
test the system clock but the log throttling mechanism itself, using a
simulated clock is a better option.

This also has the effect of making the test very fast as it no longer
needs to sleep to wait for the log throttling to stop.