Fix deadlock in server address updates
There was a deadlock in unit tests between the MainWorker thread and the
mxs::thread_pool threads. The name resolution of the server address was
done in the thread pool and after that, a call to WorkerGlobal::assign()
was done. Since the assign() call also updates the value on the
MainWorker, the deadlock would occur on shutdown where the MainWorker
was waiting for the thread pool to stop while the thread pool was
waiting for the MainWorker to update its value.

The solution to this problem is to not call assign() in the thread pool
but on the MainWorker. This comes at no real extra cost as the monitor
state was already updated on the MainWorker and the amount of messages
remains the same.