Use 95th percentile of latency in test_worker
This is a slightly more stable metric compared to an absolute limit of
50ms. Running all of the unit tests in parallel sometimes caused the
50ms difference to be exceeded by a few milliseconds. In develop, where
the number of unit tests is much higher, the test fails more often.

The 95th percentile of the longest duation that has the shortest vector
ends up being roughly the 10th value in a vector of size 12. The
shortest duration with the largest vector has around 36 values of which
the 95th percentile index is 33. This rules out extreme outliers where
the process could've been scheduled out but it still keeps the detection
relatively accurate.