MXS-4758: Use shared_ptr<AuthenticationData>
Storing the AuthenticationData in a shared_ptr and storing a copy of
that inside BackendAuthData guarantees that the AuthenticationData
instance remains the same for the duration of the initial authentication
phase. The same applies to the COM_CHANGE_USER: since it's also
completed asynchronously, updating the copy before assinging
m_authenticator again guarantees that the instance remains the same.