MXS-4783: Send generated responses via lcall()
Directly queuing the response in the set_response() function solves two
problems: it allows multiple responses to be delivered if multiple calls
to routeQuery() would be made and it removes the need to check whether a
result needs to be delivered.

The first problem can occur with multiple layers of readwritesplit
services that both generate the SET commands that enable GTID
tracking. If a cache filter has the response to those SET commands, each
generated SET command would end up generating a response which caused a
debug assertion to be hit.

Solving the second problem simplifies the code as the result will be
delivered, if possible, by the normal lcall mechanism. This will make
responses set with set_response() appear exactly as if they originated
from the network.