MXS-4849 Make fit into view functions reusable
The `DagGraph` component does not currently auto-fit the graph to
the view. To export the graph with all nodes visible, the graph board
should automatically pan and zoom to fit them. The functions to do so
are already implemented in `ErdWke`.

This makes `getGraphExtent` and `calcFitZoom` functions reusable.
This also refactors `getGraphExtent` to improve efficiency by reducing
the number of iterations over nodes.