MXS-4783: Create correct mxs::Reply in set_response()
This way the state of the mxs::Reply is at least mostly the same when
it's generated by a filter or a router. The mariadb::make_reply() would
ideally use the same packet parsing code that is used in the backend
protocol but that is not as trivial to implement as the rudimentary
processing that's currently done in the function.