MXS-4783: Add API error stress test
The test case creates permutations of the configuration by adding,
removing and linking services to each other. Combined with the "random"
API errors that are returned, this should reveal any memory leaks and
other problems in the API.

The test is deterministic in the order which the services are created,
linked and destroyed but the exact ordering of the SQL queries and the
generated API errors is not deterministic.

The initial version of this test creates readwritesplit and
readconnroute services with some of the simpler filters added to them.