MXS-5152: Handle shard map update failures
This commit modifies the updating to fail gracefully in case the update
could not be sent. Previously this caused the update to be assumed to
start but it never did as no response would ever be created.

The ID of the shard map updater is now stored per shard. In the case
where there are multiple sessions who are waiting for a shard map
update, it's useful to know which session is doing the update. This is
especially useful when the updating session seems to be stuck somewhere.