MXS-5158: Close session on failed database mapping
If all of the backends were unable to produce a result with any rows,
the database mapping is considered failed. The session could try
updating it again but this may lead to a permanent loop where the
database mapping never stops for the given user.

A straightforward solution to this is to cancel the update and kill the
session if such a situation occurs. The cases in which this would happen
are so rare that this should be an acceptable compromise when it comes
to simplicity of implementation and robustness of behavior.