Assert that MariaDB connection is open
Added debug assertions to all cases where the functions are used with an
object that is not open. These cases are an indication of a programming
error and having debug assertions for them should help avoid these
situations.

These have been recently visible in the system tests where the Galera
cluster initialization is taking place. The debug assertion should
generate a stacktrace which should show where the error originates from.