Make rworker_basics faster and more stable
The sleeps in the test can be replaced with busy polling. This speeds up
the test by about 60 seconds.

The test also now waits for the expected number of threads to appear and
for the termination to stop before proceeding with other test
cases. With the removal of the sleeps from the test, the test started to
fail repeatedly as the termination from one of the smoke tests was still
ongoing while the stress test started. With the new code, this is no
longer a problem.

Made the output a little bit less verbose by only printing the thread
states if they changed.