MXS-4783: Routables in set_response() must be valid
The downstream Routable in set_response() might not be valid when the
lcall() resolves if one routeQuery call fails but a subsequent one is
served from a cache filter's cache. However in general, the validity of
both Routables should be validated even if this isn't strictly necessary
in all situations.