MXS-4783: Use random seed in sts_combinations
Using a random seed should be a convenient way to generate many
configuration permutations without having to run the test for a long
time. As as long as the seed is printed at the start of the test, the
random engine state can be reproduced later on.

This also solves the problem of an ever decreasing success rate for
queries. As the depth of the routeQuery chain increases, the likelihood
of an error being generated increases. By randomizing the order of the
modules being created, they all should eventually be tested equally.