MXS-5159 Clarify client address handling
The client address was stored in multiple locations in both
binary and string formats. Thus, when a proxy protocol header
was received, multiple variables had to be modified. This patch
removes the extra address fields and explains in comments whether
a field is affected by proxy protocol and if it's normalized.