MXS-4783: Use trace logging in sts_combinations
Given the (newly added) random behavior of the test, being able to
capture the trace log output over several runs becomes very
important. Storing all output with log_info is tricky as the log volume
created by the test case is very high. The files are stored in /dev/shm
which should always be a tmpfs mount. Storing the files in /tmp wouldn't
always work as, for example in Rocky Linux 8 VMs, it may be a regular
disk.

If the test does fail, the random seed is printed again. This makes it
easier to see it when there's a lot of output from the test.