MXS-5157 Install `sql-limiter` package
The mentioned package can be used for both splitting SQL into
statements and injecting/enforcing the LIMIT clause. However, it
currently doesn't support injecting `OFFSET`. Therefore, this installs
the forked version of it while waiting for the pull requests to be
approved by the package's maintainer.

[Pull Request](https://github.com/sqlpad/sql-limiter/pull/26)

This also uninstalls `dbgate-query-splitter` as `sql-limiter` can
also be used for splitting statements, although it doesn't support
compound statements. There should be a separate UI for handling
compound statements to reduce the complexity that the SQL text editor
has to handle. ´