MXS-5166 Add `addStatementInfo` helper
The "Results", "Data Preview", and "Processlist" tabs will be adjusted
later to consistently display the executed statement and duration timer
info. To achieve this, they all require the `statement` property within
the result set. Therefore, this adds a reusable helper function for adding
this property to the result set.

This also adds `request_sent_time` and `total_duration` for `process_list`
so that the `DurationTimer` can be shown for the "Processlist" tab
as well.