MXS-5166 Add `reload` feature
The user can now reload the resultset with a customized limit value.

This also renames the label for `query_row_limit` input from `Limit`
to `Default limit` as it indicates the default maximum number of rows
to return from `SELECT` statements. Meanwhile, the label for the one
in the `DataTableToolbar` will be `Limit`.

The limit input in `DataTableToolbar` is kept in memory and is not
related to the `Default limit`.