MXS-5166 Remove `query_row_offset` from preferences state module
In hindsight, there is no need to persist this value in the
preferences state module. The `OFFSET` value is intended to be
dynamic and user-configurable during a single session but doesn't
need to persist beyond that session. Storing it in the preferences
state module, which is designed for persistent user settings, would
be unnecessary and misleading.