MXS-5166 Use `mode` for `enforceLimitOffset`
From the UX perspective, if the user manually specifies a
limit in their queries, the default limit should be ignored.
It should only be injected if the query does not include it.
Therefore, a `mode` parameter would provide more robust control
compared to the current `shouldReplace` parameter.

The forked version of `sql-limiter` is now refactored to use
`mode` parameter. So, it needs to be reinstalled.

This also renames `limitNumber, offsetNumber` to `limit, offset`.