MXS-5166 Improve the UX for limit and offset inputs in the table toolbar
This improves the input validation and position of inputs.

The limit injection is done with the following syntax:
`LIMIT row_count OFFSET offset`, so it would be more reasonable
to place the limit input before the offset input.

The current validation behavior for treating empty values is causing
discomfort since the default offset input is 0. Some users tend to
delete the existing value before entering a new one, which results
in a validation error for having an empty value. If the values
are empty, the default values will be used.