MXS-5166 Rename `request_sent_time` and  refactor `total_duration`
Keeping two different time units makes it difficult to reason about the data.
The `request_sent_time` is the start time when the query request begins,
measured in milliseconds, while the `total_duration` is the elapsed time,
measured in seconds.

To keep the data clear and improve readability, this renames
`request_sent_time` to `start_time` and refactors `total_duration`
to `end_time`.