Fix `vue/no-side-effects-in-computed-properties`
With the new version of eslint, that rule is now turned on by default.

"This rule is aimed at preventing the code which makes side effects
in computed properties and functions. It is considered a very bad
practice to introduce side effects inside computed properties and
functions. It makes the code not predictable and hard to understand."

To address side effects in some computed properties, this adds
`useCountUniqueValues` function to retrieve unique values based
on the provided field.