MXS-4886 Rename `DdlEditor` to `TableStructureEditor`
The component is designed for creating or altering tables, but its
current name is too generic. So this renames it to `TableStructureEditor`.

The name `DdlEditor` will be reserved for later use, which is for
creating or altering functions, views, and stored procedures.

This also renames `DDL_EDITOR_SPECS` to `TABLE_STRUCTURE_SPECS`.