Rename maxgui constants
It's not easy to immediately recognize whether a variable is an
array or an object map with the current naming convention for
constant variables. For example, `NODE_TYPES` can be misunderstood
as an array, but its value is actually an object map. Using a more
descriptive naming convention for constant variables can help
clarify their data structures.

Object maps will now have `_MAP` as a suffix to make their data
structure more immediately recognizable.