MXS-5183 Allow previewing table data or its details in a new query tab
If the current query tab is used for altering a table or viewing
insights data, the options to preview table data and its details
are disabled without showing any reasonable explanation to the
user. Instead of being disabled, these options should automatically
open a new query tab to show the result data if the current query
tab is not the default text editor.