MXS-5185 Manually watch and unwatch `isTabMoveFocus`
When `SqlEditor` is kept alive, its watcher remains active even if
the component is in an inactive state. This leads to unexpected
updates to the `toggleTabFocusMode` editor action. To address this,
the watcher for the `isTabMoveFocus` prop will be manually added
and removed when the component is configured to be kept alive.