MXS-5190 & MXS-5193: Treat multi-statements as writes
The fact that the query itself wasn't routed to all nodes in 21.06 and
older versions is because of the target_node variable that
readwritesplit still used. The call to lock_to_master() sets the
variable to the current master which causes the subsequent
is_locked_to_master() call to return true which forces the query target
to be TARGET_MASTER.

The reason why it was added to the history was because the client
protocol always returned true from lock_to_master() and false from
is_locked_to_master(). Thus the same code that readwritesplit ended up
using was never reached by the client protocol which caused the query
target to be TARGET_ALL.

By always treating multi-statement commands as writes, the
classification of them is always the same regardless of the
configuration and where the QueryClassifier is used. Additionally, the
check for whether the protocol has enabled support for multi-statements
shouldn't have affected how the parsing of the query works. There is a
performance reason for this but it comes at the cost of additional
complexity that in later releases has been optimized away by storing the
multi-statement information in the query classifier cache.